1124 binarySearch update

+++ back-end/src/main/java/com/ajoubooking/demo/dto/CallNumberDto.java
@@ -3,9 +3,11 @@ package com.ajoubooking.demo.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 public class CallNumberDto {
-    private Float classificationNumber;
+    private BigDecimal classificationNumber;
     private String authorSymbol;
 }
